Added settings page. Fixed custom keybind problems. Desktop only.
Added the settings page toggles for the chord and vim mode display.
Fixed the custom keybinds not resetting when completing. The
"vim-command-done" event seems to be not registering them. Maybe a
problem with initialization of the keys?
Changed the plugin to be desktop only. It's unlikely that someone could
use this on mobile, and cmEditor is deprecated and will only work on
desktop for now.